+++ crates/nu_plugin_custom_values/src/drop_check.rs
@@ -1,14 +1,19 @@
-use nu_protocol::{record, CustomValue, ShellError, Span, Value};
+use nu_plugin::{EngineInterface, EvaluatedCall, LabeledError, SimplePluginCommand};
+use nu_protocol::{
+    record, Category, CustomValue, PluginSignature, ShellError, Span, SyntaxShape, Value,
+};
 use serde::{Deserialize, Serialize};
 
+use crate::CustomValuePlugin;
+
 #[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct DropCheck {
+pub struct DropCheckValue {
     pub(crate) msg: String,
 }
 
-impl DropCheck {
-    pub(crate) fn new(msg: String) -> DropCheck {
-        DropCheck { msg }
+impl DropCheckValue {
+    pub(crate) fn new(msg: String) -> DropCheckValue {
+        DropCheckValue { msg }
     }
 
     pub(crate) fn into_value(self, span: Span) -> Value {
@@ -16,18 +21,18 @@ impl DropCheck {
     }
 
     pub(crate) fn notify(&self) {
-        eprintln!("DropCheck was dropped: {}", self.msg);
+        eprintln!("DropCheckValue was dropped: {}", self.msg);
     }
 }
 
 #[typetag::serde]
-impl CustomValue for DropCheck {
+impl CustomValue for DropCheckValue {
     fn clone_value(&self, span: Span) -> Value {
         self.clone().into_value(span)
     }
 
     fn value_string(&self) -> String {
-        "DropCheck".into()
+        "DropCheckValue".into()
     }
 
     fn to_base_value(&self, span: Span) -> Result<Value, ShellError> {
@@ -46,5 +51,28 @@ impl CustomValue for DropCheck {
     fn notify_plugin_on_drop(&self) -> bool {
         // This is what causes Nushell to let us know when the value is dropped
         true
+    }
+}
+
+pub struct DropCheck;
+
+impl SimplePluginCommand for DropCheck {
+    type Plugin = CustomValuePlugin;
+
+    fn signature(&self) -> nu_protocol::PluginSignature {
+        PluginSignature::build("custom-value drop-check")
+            .usage("Generates a custom value that prints a message when dropped")
+            .required("msg", SyntaxShape::String, "the message to print on drop")
+            .category(Category::Experimental)
+    }
+
+    fn run(
+        &self,
+        _plugin: &Self::Plugin,
+        _engine: &EngineInterface,
+        call: &EvaluatedCall,
+        _input: &Value,
+    ) -> Result<Value, LabeledError> {
+        Ok(DropCheckValue::new(call.req(0)?).into_value(call.head))
     }
 }
