+++ crates/nu_plugin_example/src/commands/mod.rs
@@ -1,0 +1,13 @@
+mod nu_example_1;
+mod nu_example_2;
+mod nu_example_3;
+mod nu_example_config;
+mod nu_example_disable_gc;
+mod nu_example_env;
+
+pub use nu_example_1::NuExample1;
+pub use nu_example_2::NuExample2;
+pub use nu_example_3::NuExample3;
+pub use nu_example_config::NuExampleConfig;
+pub use nu_example_disable_gc::NuExampleDisableGc;
+pub use nu_example_env::NuExampleEnv;
