+++ crates/nu_plugin_formats/src/from/ics.rs
@@ -1,52 +1,76 @@
 use ical::parser::ical::component::*;
 use ical::property::Property;
 use indexmap::map::IndexMap;
-use nu_plugin::{EvaluatedCall, LabeledError};
-use nu_protocol::{record, PluginExample, ShellError, Span, Value};
+use nu_plugin::{EngineInterface, EvaluatedCall, LabeledError, SimplePluginCommand};
+use nu_protocol::{
+    record, Category, PluginExample, PluginSignature, ShellError, Span, Type, Value,
+};
 use std::io::BufReader;
+
+use crate::FromCmds;
 
 pub const CMD_NAME: &str = "from ics";
 
-pub fn from_ics_call(call: &EvaluatedCall, input: &Value) -> Result<Value, LabeledError> {
-    let span = input.span();
-    let input_string = input.coerce_str()?;
-    let head = call.head;
+pub struct FromIcs;
 
-    let input_string = input_string
-        .lines()
-        .enumerate()
-        .map(|(i, x)| {
-            if i == 0 {
-                x.trim().to_string()
-            } else if x.len() > 1 && (x.starts_with(' ') || x.starts_with('\t')) {
-                x[1..].trim_end().to_string()
-            } else {
-                format!("\n{}", x.trim())
-            }
-        })
-        .collect::<String>();
+impl SimplePluginCommand for FromIcs {
+    type Plugin = FromCmds;
 
-    let input_bytes = input_string.as_bytes();
-    let buf_reader = BufReader::new(input_bytes);
-    let parser = ical::IcalParser::new(buf_reader);
-
-    let mut output = vec![];
-
-    for calendar in parser {
-        match calendar {
-            Ok(c) => output.push(calendar_to_value(c, head)),
-            Err(e) => output.push(Value::error(
-                ShellError::UnsupportedInput {
-                    msg: format!("input cannot be parsed as .ics ({e})"),
-                    input: "value originates from here".into(),
-                    msg_span: head,
-                    input_span: span,
-                },
-                span,
-            )),
-        }
+    fn signature(&self) -> nu_protocol::PluginSignature {
+        PluginSignature::build(CMD_NAME)
+            .input_output_types(vec![(Type::String, Type::Table(vec![]))])
+            .usage("Parse text as .ics and create table.")
+            .plugin_examples(examples())
+            .category(Category::Formats)
     }
-    Ok(Value::list(output, head))
+
+    fn run(
+        &self,
+        _plugin: &FromCmds,
+        _engine: &EngineInterface,
+        call: &EvaluatedCall,
+        input: &Value,
+    ) -> Result<Value, LabeledError> {
+        let span = input.span();
+        let input_string = input.coerce_str()?;
+        let head = call.head;
+
+        let input_string = input_string
+            .lines()
+            .enumerate()
+            .map(|(i, x)| {
+                if i == 0 {
+                    x.trim().to_string()
+                } else if x.len() > 1 && (x.starts_with(' ') || x.starts_with('\t')) {
+                    x[1..].trim_end().to_string()
+                } else {
+                    format!("\n{}", x.trim())
+                }
+            })
+            .collect::<String>();
+
+        let input_bytes = input_string.as_bytes();
+        let buf_reader = BufReader::new(input_bytes);
+        let parser = ical::IcalParser::new(buf_reader);
+
+        let mut output = vec![];
+
+        for calendar in parser {
+            match calendar {
+                Ok(c) => output.push(calendar_to_value(c, head)),
+                Err(e) => output.push(Value::error(
+                    ShellError::UnsupportedInput {
+                        msg: format!("input cannot be parsed as .ics ({e})"),
+                        input: "value originates from here".into(),
+                        msg_span: head,
+                        input_span: span,
+                    },
+                    span,
+                )),
+            }
+        }
+        Ok(Value::list(output, head))
+    }
 }
 
 pub fn examples() -> Vec<PluginExample> {
