+++ crates/nu_plugin_gstat/src/lib.rs
@@ -2,3 +2,4 @@ mod gstat;
 mod nu;
 
 pub use gstat::GStat;
+pub use nu::GStatPlugin;
