+++ crates/nu_plugin_gstat/src/main.rs
@@ -1,6 +1,6 @@
 use nu_plugin::{serve_plugin, MsgPackSerializer};
-use nu_plugin_gstat::GStat;
+use nu_plugin_gstat::GStatPlugin;
 
 fn main() {
-    serve_plugin(&GStat::new(), MsgPackSerializer {})
+    serve_plugin(&GStatPlugin, MsgPackSerializer {})
 }
