+++ crates/nu_plugin_inc/src/lib.rs
@@ -2,3 +2,4 @@ mod inc;
 mod nu;
 
 pub use inc::Inc;
+pub use nu::IncPlugin;
