+++ crates/nu_plugin_inc/src/main.rs
@@ -1,6 +1,6 @@
 use nu_plugin::{serve_plugin, JsonSerializer};
-use nu_plugin_inc::Inc;
+use nu_plugin_inc::IncPlugin;
 
 fn main() {
-    serve_plugin(&Inc::new(), JsonSerializer {})
+    serve_plugin(&IncPlugin, JsonSerializer {})
 }
