+++ crates/nu_plugin_query/src/query.rs
@@ -1,9 +1,10 @@
-use crate::query_json::execute_json_query;
-use crate::query_web::parse_selector_params;
-use crate::query_xml::execute_xpath_query;
+use crate::query_json::QueryJson;
+use crate::query_web::QueryWeb;
+use crate::query_xml::QueryXml;
+
 use nu_engine::documentation::get_flags_section;
-use nu_plugin::{EvaluatedCall, LabeledError, Plugin};
-use nu_protocol::{PluginSignature, Spanned, Value};
+use nu_plugin::{EvaluatedCall, LabeledError, Plugin, PluginCommand, SimplePluginCommand};
+use nu_protocol::{Category, PluginSignature, Value};
 use std::fmt::Write;
 
 #[derive(Default)]
@@ -17,48 +18,50 @@ impl Query {
     pub fn usage() -> &'static str {
         "Usage: query"
     }
+}
 
-    pub fn query(
-        &self,
-        _name: &str,
-        call: &EvaluatedCall,
-        _value: &Value,
-        _path: Option<Spanned<String>>,
-    ) -> Result<Value, LabeledError> {
-        let help = get_brief_subcommand_help(&Query.signature());
-        Ok(Value::string(help, call.head))
-    }
-
-    pub fn query_json(
-        &self,
-        name: &str,
-        call: &EvaluatedCall,
-        input: &Value,
-        query: Option<Spanned<String>>,
-    ) -> Result<Value, LabeledError> {
-        execute_json_query(name, call, input, query)
-    }
-    pub fn query_web(
-        &self,
-        _name: &str,
-        call: &EvaluatedCall,
-        input: &Value,
-        _rest: Option<Spanned<String>>,
-    ) -> Result<Value, LabeledError> {
-        parse_selector_params(call, input)
-    }
-    pub fn query_xml(
-        &self,
-        name: &str,
-        call: &EvaluatedCall,
-        input: &Value,
-        query: Option<Spanned<String>>,
-    ) -> Result<Value, LabeledError> {
-        execute_xpath_query(name, call, input, query)
+impl Plugin for Query {
+    fn commands(&self) -> Vec<Box<dyn PluginCommand<Plugin = Self>>> {
+        vec![
+            Box::new(QueryCommand),
+            Box::new(QueryJson),
+            Box::new(QueryXml),
+            Box::new(QueryWeb),
+        ]
     }
 }
 
-pub fn get_brief_subcommand_help(sigs: &[PluginSignature]) -> String {
+// With no subcommand
+pub struct QueryCommand;
+
+impl SimplePluginCommand for QueryCommand {
+    type Plugin = Query;
+
+    fn signature(&self) -> PluginSignature {
+        PluginSignature::build("query")
+            .usage("Show all the query commands")
+            .category(Category::Filters)
+    }
+
+    fn run(
+        &self,
+        _plugin: &Query,
+        _engine: &nu_plugin::EngineInterface,
+        call: &EvaluatedCall,
+        _input: &Value,
+    ) -> Result<Value, LabeledError> {
+        let help = get_brief_subcommand_help();
+        Ok(Value::string(help, call.head))
+    }
+}
+
+pub fn get_brief_subcommand_help() -> String {
+    let sigs: Vec<_> = Query
+        .commands()
+        .into_iter()
+        .map(|cmd| cmd.signature())
+        .collect();
+
     let mut help = String::new();
     let _ = write!(help, "{}\n\n", sigs[0].sig.usage);
     let _ = write!(help, "Usage:\n  > {}\n\n", sigs[0].sig.name);
