+++ crates/nu_plugin_query/src/query_web.rs
@@ -1,7 +1,75 @@
-use crate::web_tables::WebTable;
-use nu_plugin::{EvaluatedCall, LabeledError};
-use nu_protocol::{Record, Span, Value};
+use crate::{web_tables::WebTable, Query};
+use nu_plugin::{EngineInterface, EvaluatedCall, LabeledError, SimplePluginCommand};
+use nu_protocol::{Category, PluginExample, PluginSignature, Record, Span, SyntaxShape, Value};
 use scraper::{Html, Selector as ScraperSelector};
+
+pub struct QueryWeb;
+
+impl SimplePluginCommand for QueryWeb {
+    type Plugin = Query;
+
+    fn signature(&self) -> PluginSignature {
+        PluginSignature::build("query web")
+            .usage("execute selector query on html/web")
+            .named("query", SyntaxShape::String, "selector query", Some('q'))
+            .switch("as-html", "return the query output as html", Some('m'))
+            .plugin_examples(web_examples())
+            .named(
+                "attribute",
+                SyntaxShape::String,
+                "downselect based on the given attribute",
+                Some('a'),
+            )
+            .named(
+                "as-table",
+                SyntaxShape::List(Box::new(SyntaxShape::String)),
+                "find table based on column header list",
+                Some('t'),
+            )
+            .switch(
+                "inspect",
+                "run in inspect mode to provide more information for determining column headers",
+                Some('i'),
+            )
+            .category(Category::Network)
+    }
+
+    fn run(
+        &self,
+        _plugin: &Query,
+        _engine: &EngineInterface,
+        call: &EvaluatedCall,
+        input: &Value,
+    ) -> Result<Value, LabeledError> {
+        parse_selector_params(call, input)
+    }
+}
+
+pub fn web_examples() -> Vec<PluginExample> {
+    vec![
+        PluginExample {
+            example: "http get https://phoronix.com | query web --query 'header' | flatten".into(),
+            description: "Retrieve all `<header>` elements from phoronix.com website".into(),
+            result: None,
+        },
+        PluginExample {
+            example: "http get https://en.wikipedia.org/wiki/List_of_cities_in_India_by_population |
+        query web --as-table [City 'Population(2011)[3]' 'Population(2001)[3][a]' 'State or unionterritory' 'Ref']".into(),
+            description: "Retrieve a html table from Wikipedia and parse it into a nushell table using table headers as guides".into(),
+            result: None
+        },
+        PluginExample {
+            example: "http get https://www.nushell.sh | query web --query 'h2, h2 + p' | each {str join} | group 2 | each {rotate --ccw tagline description} | flatten".into(),
+            description: "Pass multiple css selectors to extract several elements within single query, group the query results together and rotate them to create a table".into(),
+            result: None,
+        },
+        PluginExample {
+            example: "http get https://example.org | query web --query a --attribute href".into(),
+            description: "Retrieve a specific html attribute instead of the default text".into(),
+            result: None,
+        }
+    ]
+}
 
 pub struct Selector {
     pub query: String,
