+++ crates/nu_plugin_stream_example/src/commands/collect_external.rs
@@ -1,0 +1,51 @@
+use nu_plugin::{EngineInterface, EvaluatedCall, LabeledError, PluginCommand};
+use nu_protocol::{Category, PipelineData, PluginExample, PluginSignature, RawStream, Type, Value};
+
+use crate::StreamExample;
+
+/// `<list<string>> | stream_example collect-external`
+pub struct CollectExternal;
+
+impl PluginCommand for CollectExternal {
+    type Plugin = StreamExample;
+
+    fn signature(&self) -> PluginSignature {
+        PluginSignature::build("stream_example collect-external")
+            .usage("Example transformer to raw external stream")
+            .search_terms(vec!["example".into()])
+            .input_output_types(vec![
+                (Type::List(Type::String.into()), Type::String),
+                (Type::List(Type::Binary.into()), Type::Binary),
+            ])
+            .plugin_examples(vec![PluginExample {
+                example: "[a b] | stream_example collect-external".into(),
+                description: "collect strings into one stream".into(),
+                result: Some(Value::test_string("ab")),
+            }])
+            .category(Category::Experimental)
+    }
+
+    fn run(
+        &self,
+        _plugin: &StreamExample,
+        _engine: &EngineInterface,
+        call: &EvaluatedCall,
+        input: PipelineData,
+    ) -> Result<PipelineData, LabeledError> {
+        let stream = input.into_iter().map(|value| {
+            value
+                .as_str()
+                .map(|str| str.as_bytes())
+                .or_else(|_| value.as_binary())
+                .map(|bin| bin.to_vec())
+        });
+        Ok(PipelineData::ExternalStream {
+            stdout: Some(RawStream::new(Box::new(stream), None, call.head, None)),
+            stderr: None,
+            exit_code: None,
+            span: call.head,
+            metadata: None,
+            trim_end_newline: false,
+        })
+    }
+}
