+++ crates/nu_plugin_stream_example/src/commands/mod.rs
@@ -1,0 +1,11 @@
+mod collect_external;
+mod for_each;
+mod generate;
+mod seq;
+mod sum;
+
+pub use collect_external::CollectExternal;
+pub use for_each::ForEach;
+pub use generate::Generate;
+pub use seq::Seq;
+pub use sum::Sum;
